# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service for interacting with Vodacom's DRC MPESA SOAP Integrated Payment Gateway.",
        "title": "DRC MPESA Proxy REST API",
        "termsOfService": "https://blog.tralahm.com",
        "contact": {
            "name": "API Support",
            "email": "briantralah@gmail.com"
        },
        "license": {
            "name": "GNU GENERAL PUBLIC LICENSE",
            "url": "http://www.gnu.org/licenses/"
        },
        "version": "1.0"
    },
    "host": "ipg.betmondenge.com",
    "basePath": "/",
    "paths": {
        "/api/v1/b2c": {
            "post": {
                "description": "Initiate B2C Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "b2c"
                ],
                "summary": "Initiate Business to Customer Transaction",
                "parameters": [
                    {
                        "description": "B2C",
                        "name": "b2c",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.B2C"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.B2CResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/b2c_callback": {
            "post": {
                "description": "Handle CallBack for a B2C Transaction and POST JSON callback to client callback url",
                "consumes": [
                    "text/xml"
                ],
                "produces": [
                    "text/xml"
                ],
                "tags": [
                    "b2c"
                ],
                "summary": "Handle CallBack for a Customer to Business Transaction",
                "parameters": [
                    {
                        "description": "B2C",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.B2CCallbackEnvelope"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.B2CAcknowledgement"
                        }
                    }
                }
            }
        },
        "/api/v1/c2b": {
            "post": {
                "description": "Initiate C2B Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "c2b"
                ],
                "summary": "Initiate Customer to Business Transaction",
                "parameters": [
                    {
                        "description": "C2B",
                        "name": "c2b",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.C2B"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.C2BResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/c2b_callback": {
            "post": {
                "description": "Handle CallBack for a C2B Transaction and POST JSON callback to client callback url",
                "consumes": [
                    "text/xml"
                ],
                "produces": [
                    "text/xml"
                ],
                "tags": [
                    "c2b"
                ],
                "summary": "Handle CallBack for a Customer to Business Transaction",
                "parameters": [
                    {
                        "description": "C2B",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.C2BCallbackEnvelope"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.C2BAcknowledgement"
                        }
                    }
                }
            }
        },
...
        "/api/v1/login": {
            "post": {
                "description": "Login to the MPESA Ipg with the credentials and return JSON response.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "login"
                ],
                "summary": "Authenticate against the Remote IPG",
                "parameters": [
                    {
                        "description": "Login",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Login"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/main.LoginResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.B2C": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "callBackChannel": {
                    "type": "string"
                },
                "callBackDestination": {
                    "type": "string"
                },
                "commandID": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "customerMSISDN": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "serviceProviderName": {
                    "type": "string"
                },
                "shortcode": {
                    "type": "string"
                },
                "thirdPartyReference": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "transactionDateTime": {
                    "type": "string"
                }
            }
        },
        "main.B2CAcknowledgement": {
            "type": "object",
            "properties": {
                "dataItem": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "text": {
                                "type": "string"
                            },
                            "type": {
                                "type": "string"
                            },
                            "value": {
                                "type": "string"
                            }
                        }
                    }
                },
                "text": {
                    "type": "string"
                },
                "xmlname": {
                    "$ref": "#/definitions/main.Name"
                }
            }
        },
        "main.B2CCallbackEnvelope": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "properties": {
                        "getGenericResult": {
                            "type": "object",
                            "properties": {
                                "request": {
                                    "type": "object",
                                    "properties": {
                                        "dataItem": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "text": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "value": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "text": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "text": {
                                    "type": "string"
                                }
                            }
                        },
                        "text": {
                            "type": "string"
                        }
                    }
                },
                "gen": {
                    "type": "string"
                },
                "header": {
                    "type": "object",
                    "properties": {
                        "eventID": {
                            "type": "string"
                        },
                        "text": {
                            "type": "string"
                        }
                    }
                },
                "soap": {
                    "type": "string"
                },
                "soapenv": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "xmlname": {
                    "$ref": "#/definitions/main.Name"
                }
            }
        },
        "main.B2CResponse": {
            "type": "object",
            "properties": {
                "Amount": {
                    "type": "string"
                },
                "CallBackChannel": {
                    "type": "string"
                },
                "CallBackDestination": {
                    "type": "string"
                },
                "CommandID": {
                    "type": "string"
                },
                "Currency": {
                    "type": "string"
                },
                "CustomerMSISDN": {
                    "type": "string"
                },
                "InsightReference": {
                    "type": "string"
                },
                "Language": {
                    "type": "string"
                },
                "ResponseCode": {
                    "type": "string"
                },
                "ServiceProviderName": {
                    "type": "string"
                },
                "Shortcode": {
                    "type": "string"
                },
                "ThirdPartyReference": {
                    "type": "string"
                },
                "TransactionDateTime": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "transactionID": {
                    "type": "string"
                }
            }
        },
        "main.C2B": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "callBackChannel": {
                    "type": "string"
                },
                "callBackDestination": {
                    "type": "string"
                },
                "commandID": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "customerMSISDN": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "initials": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "serviceProviderCode": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "thirdPartyReference": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "main.C2BResponse": {
            "type": "object",
            "properties": {
                "Amount": {
                    "type": "string"
                },
                "CallBackChannel": {
                    "type": "string"
                },
                "CallBackDestination": {
                    "type": "string"
                },
                "CommandID": {
                    "type": "string"
                },
                "Currency": {
                    "type": "string"
                },
                "CustomerMSISDN": {
                    "type": "string"
                },
                "Date": {
                    "type": "string"
                },
                "Initials": {
                    "type": "string"
                },
                "InsightReference": {
                    "type": "string"
                },
                "Language": {
                    "type": "string"
                },
                "ResponseCode": {
                    "type": "string"
                },
                "ServiceProviderCode": {
                    "type": "string"
                },
                "Surname": {
                    "type": "string"
                },
                "ThirdPartyReference": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "event_id": {
                    "type": "string"
                },
                "transactionID": {
                    "type": "string"
                }
            }
        },
...
    }
}